refactor: 抽離種子資料工具模組
